cmd/sctool: move base url code from config/server and get rid of config/server dependency completely
