+++ pkg/config/server/server_test.go
@@ -104,61 +104,3 @@ func TestDefaultConfig(t *testing.T) {
 		t.Fatal(diff)
 	}
 }
-
-func TestBaseURL(t *testing.T) {
-	table := []struct {
-		Name   string
-		Config server.Config
-		Golden string
-	}{
-		{
-			Name:   "empty configuration",
-			Config: server.Config{},
-			Golden: "",
-		},
-		{
-			Name:   "HTTP",
-			Config: server.Config{HTTP: "127.0.0.1:12345"},
-			Golden: "http://127.0.0.1:12345/api/v1",
-		},
-		{
-			Name:   "HTTPS",
-			Config: server.Config{HTTPS: "127.0.0.1:54321"},
-			Golden: "https://127.0.0.1:54321/api/v1",
-		},
-		{
-			Name:   "HTTP override",
-			Config: server.Config{HTTP: "127.0.0.1:12345", HTTPS: "127.0.0.2:54321"},
-			Golden: "http://127.0.0.1:12345/api/v1",
-		},
-		{
-			Name:   "HTTP override on all interfaces",
-			Config: server.Config{HTTP: "0.0.0.0:12345", HTTPS: "127.0.0.1:54321"},
-			Golden: "http://127.0.0.1:12345/api/v1",
-		},
-		{
-			Name:   "HTTP empty host",
-			Config: server.Config{HTTP: ":12345"},
-			Golden: "http://127.0.0.1:12345/api/v1",
-		},
-		{
-			Name:   "IPV6",
-			Config: server.Config{HTTP: "[::1]:12345"},
-			Golden: "http://[::1]:12345/api/v1",
-		},
-		{
-			Name:   "IPV6 on all interfaces",
-			Config: server.Config{HTTPS: "[::0]:54321"},
-			Golden: "https://[::1]:54321/api/v1",
-		},
-	}
-
-	for i := range table {
-		test := table[i]
-		t.Run(test.Name, func(t *testing.T) {
-			if diff := cmp.Diff(test.Golden, test.Config.BaseURL()); diff != "" {
-				t.Error(diff)
-			}
-		})
-	}
-}
